Apply decision rules to time series and assess emerging status

This function defines and applies some decision rules to assess emerging status at a specific time.

apply_decision_rules(
  df,
  y_var = "ncells",
  eval_year,
  year = "year",
  taxonKey = "taxonKey"
)

Arguments

df

df. A dataframe containing temporal data of one or more taxa. The column with taxa can be of class character, numeric or integers.

y_var

character. Name of column of df containing variable to model. It has to be passed as string, e.g. "occurrences".

eval_year

numeric. Temporal value at which emerging status has to be evaluated. eval_year should be present in timeseries of at least one taxon.

year

character. Name of column of df containing temporal values. It has to be passed as string, e.g. "time". Default: "year".

taxonKey

character. Name of column of df containing taxon IDs. It has to be passed as string, e.g. "taxon". Default: "taxonKey".

Value

df. A dataframe (tibble) containing emerging status. Columns:

year: column containing temporal values. Column name equal to value of argument year. Column itself is equal to value of argument eval_year. So, if you apply decision rules on years 2018 (eval_year = 2018), you will get 2018 in this column.

em_status: numeric. Emerging status, an integer between 0 and 3, based on output of decision rules (next columns).

dr_1: logical. Output of decision rule 1 answers to the question: does the time series contain only one positive value at evaluation year?

dr_2: logical. Output of decision rule 2 answers to the question: is value at evaluation year above median value?

dr_3: logical. Output of decision rule 3 answers to the question: does the time series contains only zeros in the five years before eval_year?

dr_4: logical. Output of decision rule 4 answers to the question: is the value in column y_var the maximum ever observed up to eval_year?

Examples

if (FALSE) {
df <- tibble(
  taxonID = c(rep(1008955, 10), rep(2493598, 3)),
  y = c(seq(2009, 2018), seq(2016, 2018)),
  obs = c(1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 3, 0)
)
apply_decision_rules(df,
  eval_year = c(2016, 2017, 2018),
  y_var = "obs",
  taxonKey = "taxonID",
  year = y
)
}