Get taxa information from GBIF

This function retrieves taxa information from GBIF. It is a higher level function built on rgbif functions `name_usage()` and `name_lookup()`.

gbif_get_taxa(taxon_keys = NULL, checklist_keys = NULL,
  origin = NULL, limit = NULL)

Arguments

taxon_keys
(single numeric or character or a vector) a single key or a vector of keys. Not to use together with `checklist_keys`.

checklist_keys
(single character or a vector) a datasetKey (character) or a vector of datasetkeys. Not to use together with `checklist_keys`.

origin
(single character or a vector) filter by origin. It can take many inputs, and treated as OR (e.g., a or b or c) To be used only in combination with `checklist_keys`. Ignored otherwise.

limit
With taxon_keys: limit number of taxa. With checklist_keys: limit number of taxa per each dataset. A warning is given if limit is higher than the length of taxon_keys or number of records in the checklist_keys (if string) or any of the checklist_keys (if vector)

Value

A data.frame with all returned attributes for any taxa

Examples

# A single numeric taxon_keys
gbif_get_taxa(taxon_keys = 1)

#> [1] "All 1 taxa match GBIF Backbone."

#> # A tibble: 1 x 23
#>     key scientificName nubKey nameKey taxonID sourceTaxonKey kingdom kingdomKey
#>   <int> <chr>           <int>   <int> <chr>            <int> <chr>        <int>
#> 1     1 Animalia            1  1.85e7 gbif:1       144299179 Animal~          1
#> # ... with 15 more variables: datasetKey <chr>, constituentKey <chr>,
#> #   canonicalName <chr>, authorship <chr>, nameType <chr>, rank <chr>,
#> #   origin <chr>, taxonomicStatus <chr>, nomenclaturalStatus <lgl>,
#> #   remarks <chr>, numDescendants <int>, lastCrawled <chr>,
#> #   lastInterpreted <chr>, issues <chr>, synonym <lgl>

# A single character taxon_keys
gbif_get_taxa(taxon_keys = "1")

#> [1] "All 1 taxa match GBIF Backbone."

#> # A tibble: 1 x 23
#>     key scientificName nubKey nameKey taxonID sourceTaxonKey kingdom kingdomKey
#>   <int> <chr>           <int>   <int> <chr>            <int> <chr>        <int>
#> 1     1 Animalia            1  1.85e7 gbif:1       144299179 Animal~          1
#> # ... with 15 more variables: datasetKey <chr>, constituentKey <chr>,
#> #   canonicalName <chr>, authorship <chr>, nameType <chr>, rank <chr>,
#> #   origin <chr>, taxonomicStatus <chr>, nomenclaturalStatus <lgl>,
#> #   remarks <chr>, numDescendants <int>, lastCrawled <chr>,
#> #   lastInterpreted <chr>, issues <chr>, synonym <lgl>

# Multiple numeric taxon_keys (vector)
gbif_get_taxa(taxon_keys = c(1,2,3,4,5,6))

#> [1] "All 6 taxa match GBIF Backbone."

#> # A tibble: 6 x 23
#>     key scientificName nubKey nameKey taxonID sourceTaxonKey kingdom kingdomKey
#>   <int> <chr>           <int>   <int> <chr>            <int> <chr>        <int>
#> 1     1 Animalia            1  1.85e7 gbif:1       144299179 Animal~          1
#> 2     2 Archaea             2  1.86e7 gbif:2       144974845 Archaea          2
#> 3     3 Bacteria            3  1.86e7 gbif:3       144974961 Bacter~          3
#> 4     4 Chromista           4  1.86e7 gbif:4       144975033 Chromi~          4
#> 5     5 Fungi               5  1.86e7 gbif:5       145000585 Fungi            5
#> 6     6 Plantae             6  1.87e7 gbif:6       145014415 Plantae          6
#> # ... with 15 more variables: datasetKey <chr>, constituentKey <chr>,
#> #   canonicalName <chr>, authorship <chr>, nameType <chr>, rank <chr>,
#> #   origin <chr>, taxonomicStatus <chr>, nomenclaturalStatus <lgl>,
#> #   remarks <chr>, numDescendants <int>, lastCrawled <chr>,
#> #   lastInterpreted <chr>, issues <chr>, synonym <lgl>

# Multiple character taxon_keys (vector)
gbif_get_taxa(taxon_keys = c("1","2","3","4","5","6"))

#> [1] "All 6 taxa match GBIF Backbone."

#> # A tibble: 6 x 23
#>     key scientificName nubKey nameKey taxonID sourceTaxonKey kingdom kingdomKey
#>   <int> <chr>           <int>   <int> <chr>            <int> <chr>        <int>
#> 1     1 Animalia            1  1.85e7 gbif:1       144299179 Animal~          1
#> 2     2 Archaea             2  1.86e7 gbif:2       144974845 Archaea          2
#> 3     3 Bacteria            3  1.86e7 gbif:3       144974961 Bacter~          3
#> 4     4 Chromista           4  1.86e7 gbif:4       144975033 Chromi~          4
#> 5     5 Fungi               5  1.86e7 gbif:5       145000585 Fungi            5
#> 6     6 Plantae             6  1.87e7 gbif:6       145014415 Plantae          6
#> # ... with 15 more variables: datasetKey <chr>, constituentKey <chr>,
#> #   canonicalName <chr>, authorship <chr>, nameType <chr>, rank <chr>,
#> #   origin <chr>, taxonomicStatus <chr>, nomenclaturalStatus <lgl>,
#> #   remarks <chr>, numDescendants <int>, lastCrawled <chr>,
#> #   lastInterpreted <chr>, issues <chr>, synonym <lgl>

# Limit number of taxa (coupled with taxon_keys)
gbif_get_taxa(taxon_keys = c(1,2,3,4,5,6), limit = 3)

#> [1] "All 3 taxa match GBIF Backbone."

#> # A tibble: 3 x 23
#>     key scientificName nubKey nameKey taxonID sourceTaxonKey kingdom kingdomKey
#>   <int> <chr>           <int>   <int> <chr>            <int> <chr>        <int>
#> 1     1 Animalia            1  1.85e7 gbif:1       144299179 Animal~          1
#> 2     2 Archaea             2  1.86e7 gbif:2       144974845 Archaea          2
#> 3     3 Bacteria            3  1.86e7 gbif:3       144974961 Bacter~          3
#> # ... with 15 more variables: datasetKey <chr>, constituentKey <chr>,
#> #   canonicalName <chr>, authorship <chr>, nameType <chr>, rank <chr>,
#> #   origin <chr>, taxonomicStatus <chr>, nomenclaturalStatus <lgl>,
#> #   remarks <chr>, numDescendants <int>, lastCrawled <chr>,
#> #   lastInterpreted <chr>, issues <chr>, synonym <lgl>

# A single checklist_keys (character)
gbif_get_taxa(checklist_keys = "b3fa7329-a002-4243-a7a7-cd066092c9a6")

#> [1] "90 taxa found of which 5 could not be matched to GBIF Backbone."

#> # A tibble: 90 x 33
#>       key scientificName datasetKey  nubKey parentKey parent kingdom genus
#>     <int> <chr>          <chr>        <int>     <int> <chr>  <chr>   <chr>
#>  1 1.19e8 Capillaria     b3fa7329-~ 9707382 152157289 Anima~ Animal~ Capi~
#>  2 1.19e8 Capillaria cy~ b3fa7329-~ 7020312 119338571 Capil~ Animal~ Capi~
#>  3 1.19e8 Capillaria (H~ b3fa7329-~ 7020317 119338571 Capil~ Animal~ Capi~
#>  4 1.19e8 Pseudocapilla~ b3fa7329-~ 9780320 152157293 Trich~ Animal~ Pseu~
#>  5 1.19e8 Pseudocapilla~ b3fa7329-~      NA 119338578 Pseud~ Animal~ Pseu~
#>  6 1.19e8 Pseudocapilla~ b3fa7329-~ 7020058 152157294 Pseud~ Animal~ Pseu~
#>  7 1.19e8 Capillostrong~ b3fa7329-~ 4629935 152157293 Trich~ Animal~ Capi~
#>  8 1.19e8 Paracapillaria b3fa7329-~ 9817653 152157293 Trich~ Animal~ Para~
#>  9 1.19e8 Capillariidae  b3fa7329-~ 3249126 152157292 Trich~ Animal~ NA   
#> 10 1.19e8 Paracapillaro~ b3fa7329-~ 7019884 152157295 Parac~ Animal~ Para~
#> # ... with 80 more rows, and 25 more variables: kingdomKey <int>,
#> #   genusKey <int>, canonicalName <chr>, nameType <chr>, taxonomicStatus <chr>,
#> #   rank <chr>, origin <chr>, numDescendants <int>, numOccurrences <int>,
#> #   taxonID <chr>, habitats <lgl>, nomenclaturalStatus <lgl>,
#> #   threatStatuses <lgl>, synonym <lgl>, species <chr>, speciesKey <int>,
#> #   authorship <chr>, phylum <chr>, order <chr>, family <chr>, phylumKey <int>,
#> #   classKey <int>, orderKey <int>, familyKey <int>, class <chr>

# Multiple checklist_keys (vector)
gbif_get_taxa(checklist_keys = c("e4746398-f7c4-47a1-a474-ae80a4f18e92",
                            "b3fa7329-a002-4243-a7a7-cd066092c9a6"))

#> [1] "245 taxa found of which 9 could not be matched to GBIF Backbone."

#> # A tibble: 245 x 33
#>       key scientificName datasetKey  nubKey parentKey parent family familyKey
#>     <int> <chr>          <chr>        <int>     <int> <chr>  <chr>      <int>
#>  1 1.14e8 Acer negundo   e4746398-~ 3189866 143920246 Acera~ Acera~ 143920246
#>  2 1.14e8 Acer rufinerve e4746398-~ 7263145 143920246 Acera~ Acera~ 143920246
#>  3 1.14e8 Carpobrotus s~ e4746398-~      NA 143920247 Aizoa~ Aizoa~ 143920247
#>  4 1.14e8 Prunus seroti~ e4746398-~ 3021850 143920248 Amygd~ Amygd~ 143920248
#>  5 1.14e8 Rhus typhina   e4746398-~ 3190538 143920249 Anaca~ Anaca~ 143920249
#>  6 1.14e8 Aix galericul~ e4746398-~ 2498388 143920250 Anati~ Anati~ 143920250
#>  7 1.14e8 Alopochen aeg~ e4746398-~ 2498252 143920250 Anati~ Anati~ 143920250
#>  8 1.14e8 Anser indicus  e4746398-~ 2498035 143920250 Anati~ Anati~ 143920250
#>  9 1.14e8 Branta canade~ e4746398-~ 5232437 143920250 Anati~ Anati~ 143920250
#> 10 1.14e8 Oxyura jamaic~ e4746398-~ 2498305 143920250 Anati~ Anati~ 143920250
#> # ... with 235 more rows, and 25 more variables: canonicalName <chr>,
#> #   authorship <chr>, nameType <chr>, taxonomicStatus <chr>, origin <chr>,
#> #   numDescendants <int>, numOccurrences <int>, habitats <lgl>,
#> #   nomenclaturalStatus <lgl>, threatStatuses <lgl>, synonym <lgl>, rank <chr>,
#> #   kingdom <chr>, genus <chr>, kingdomKey <int>, genusKey <int>, taxonID <chr>,
#> #   species <chr>, speciesKey <int>, phylum <chr>, order <chr>, phylumKey <int>,
#> #   classKey <int>, orderKey <int>, class <chr>

# Limit number of taxa (coupled with checklist_keys)
gbif_get_taxa(checklist_keys = c("e4746398-f7c4-47a1-a474-ae80a4f18e92",
                            "b3fa7329-a002-4243-a7a7-cd066092c9a6"),
         limit = 30)

#> [1] "60 taxa found of which 5 could not be matched to GBIF Backbone."

#> # A tibble: 60 x 33
#>       key scientificName datasetKey  nubKey parentKey parent family familyKey
#>     <int> <chr>          <chr>        <int>     <int> <chr>  <chr>      <int>
#>  1 1.14e8 Acer negundo   e4746398-~ 3189866 143920246 Acera~ Acera~ 143920246
#>  2 1.14e8 Acer rufinerve e4746398-~ 7263145 143920246 Acera~ Acera~ 143920246
#>  3 1.14e8 Carpobrotus s~ e4746398-~      NA 143920247 Aizoa~ Aizoa~ 143920247
#>  4 1.14e8 Prunus seroti~ e4746398-~ 3021850 143920248 Amygd~ Amygd~ 143920248
#>  5 1.14e8 Rhus typhina   e4746398-~ 3190538 143920249 Anaca~ Anaca~ 143920249
#>  6 1.14e8 Aix galericul~ e4746398-~ 2498388 143920250 Anati~ Anati~ 143920250
#>  7 1.14e8 Alopochen aeg~ e4746398-~ 2498252 143920250 Anati~ Anati~ 143920250
#>  8 1.14e8 Anser indicus  e4746398-~ 2498035 143920250 Anati~ Anati~ 143920250
#>  9 1.14e8 Branta canade~ e4746398-~ 5232437 143920250 Anati~ Anati~ 143920250
#> 10 1.14e8 Oxyura jamaic~ e4746398-~ 2498305 143920250 Anati~ Anati~ 143920250
#> # ... with 50 more rows, and 25 more variables: canonicalName <chr>,
#> #   authorship <chr>, nameType <chr>, taxonomicStatus <chr>, origin <chr>,
#> #   numDescendants <int>, numOccurrences <int>, habitats <lgl>,
#> #   nomenclaturalStatus <lgl>, threatStatuses <lgl>, synonym <lgl>,
#> #   kingdom <chr>, genus <chr>, kingdomKey <int>, genusKey <int>, rank <chr>,
#> #   taxonID <chr>, species <chr>, speciesKey <int>, phylum <chr>, order <chr>,
#> #   phylumKey <int>, classKey <int>, orderKey <int>, class <chr>

# Filter by origin
gbif_get_taxa(checklist_keys = "9ff7d317-609b-4c08-bd86-3bc404b77c42",
         origin = "source", limit = 3000)

#> Warning: Dataset contains less records than limit.

#> [1] "2531 taxa found of which 75 could not be matched to GBIF Backbone."

#> # A tibble: 2,531 x 26
#>       key scientificName datasetKey nubKey parentKey parent kingdom family
#>     <int> <chr>          <chr>       <int>     <int> <chr>  <chr>   <chr> 
#>  1 1.41e8 Nymphaea x ma~ 9ff7d317-~ 7.80e6 148437314 Nymph~ Plantae Nymph~
#>  2 1.41e8 Aucuba japoni~ 9ff7d317-~ 3.03e6 148437316 Garry~ Plantae Garry~
#>  3 1.41e8 Melia azedara~ 9ff7d317-~ 5.91e6 148437315 Melia~ Plantae Melia~
#>  4 1.41e8 Sesamum indic~ 9ff7d317-~ 3.17e6 148437317 Pedal~ Plantae Pedal~
#>  5 1.41e8 Tropaeolum ma~ 9ff7d317-~ 2.89e6 148437319 Tropa~ Plantae Tropa~
#>  6 1.41e8 Blumenbachia ~ 9ff7d317-~ 5.65e6 148437320 Loasa~ Plantae Loasa~
#>  7 1.41e8 Mentzelia alb~ 9ff7d317-~ 7.40e6 148437320 Loasa~ Plantae Loasa~
#>  8 1.41e8 Elatine alsin~ 9ff7d317-~ 5.65e6 148437321 Elati~ Plantae Elati~
#>  9 1.41e8 Scaevola aemu~ 9ff7d317-~ 5.57e6 148437322 Goode~ Plantae Goode~
#> 10 1.41e8 Tarenaya hass~ 9ff7d317-~ 3.87e6 148437323 Cleom~ Plantae Cleom~
#> # ... with 2,521 more rows, and 18 more variables: kingdomKey <int>,
#> #   familyKey <int>, canonicalName <chr>, nameType <chr>, taxonomicStatus <chr>,
#> #   origin <chr>, numDescendants <int>, numOccurrences <int>, taxonID <chr>,
#> #   habitats <chr>, nomenclaturalStatus <lgl>, threatStatuses <lgl>,
#> #   synonym <lgl>, species <chr>, speciesKey <int>, rank <chr>, genus <chr>,
#> #   genusKey <int>

gbif_get_taxa(checklist_keys = "9ff7d317-609b-4c08-bd86-3bc404b77c42",
         origin = c("source" ,"denormed_classification"), limit = 3000)

#> Warning: Dataset contains less records than limit.

#> [1] "2685 taxa found of which 75 could not be matched to GBIF Backbone."

#> # A tibble: 2,685 x 26
#>       key scientificName datasetKey nubKey parentKey parent kingdom family
#>     <int> <chr>          <chr>       <int>     <int> <chr>  <chr>   <chr> 
#>  1 1.41e8 Nymphaea x ma~ 9ff7d317-~ 7.80e6 148437314 Nymph~ Plantae Nymph~
#>  2 1.41e8 Aucuba japoni~ 9ff7d317-~ 3.03e6 148437316 Garry~ Plantae Garry~
#>  3 1.41e8 Melia azedara~ 9ff7d317-~ 5.91e6 148437315 Melia~ Plantae Melia~
#>  4 1.41e8 Sesamum indic~ 9ff7d317-~ 3.17e6 148437317 Pedal~ Plantae Pedal~
#>  5 1.41e8 Tropaeolum ma~ 9ff7d317-~ 2.89e6 148437319 Tropa~ Plantae Tropa~
#>  6 1.41e8 Blumenbachia ~ 9ff7d317-~ 5.65e6 148437320 Loasa~ Plantae Loasa~
#>  7 1.41e8 Mentzelia alb~ 9ff7d317-~ 7.40e6 148437320 Loasa~ Plantae Loasa~
#>  8 1.41e8 Elatine alsin~ 9ff7d317-~ 5.65e6 148437321 Elati~ Plantae Elati~
#>  9 1.41e8 Scaevola aemu~ 9ff7d317-~ 5.57e6 148437322 Goode~ Plantae Goode~
#> 10 1.41e8 Tarenaya hass~ 9ff7d317-~ 3.87e6 148437323 Cleom~ Plantae Cleom~
#> # ... with 2,675 more rows, and 18 more variables: kingdomKey <int>,
#> #   familyKey <int>, canonicalName <chr>, nameType <chr>, taxonomicStatus <chr>,
#> #   origin <chr>, numDescendants <int>, numOccurrences <int>, taxonID <chr>,
#> #   habitats <chr>, nomenclaturalStatus <lgl>, threatStatuses <lgl>,
#> #   synonym <lgl>, species <chr>, speciesKey <int>, rank <chr>, genus <chr>,
#> #   genusKey <int>