Compare desired distribution information with actual one.

This function compares GBIF distribution information based on a single taxon key with user requests and returns a logical (TRUE or FALSE). Comparison is case insensitive. User properties for each term ar treated as OR. It is a function built on rgbif function `name_usage()`.

gbif_has_distribution(taxon_key, ...)

Arguments

taxon_key
(single numeric or character) a single taxon key.

...
one or more GBIF distribution properties and related values. Up to now it supports the following properties: country (and its synoynym: countryCode), status (and its synonym: occurrenceStatus) and establishmentMeans.

Value

a logical, TRUE or FALSE.

Examples

# numeric taxonKey, atomic parameters
gbif_has_distribution(145953242,
                      country = "BE",
                      status = "PRESENT",
                      establishmentMeans = "INTRODUCED")

#> [1] TRUE

# character taxonKey, distribution properties as vectors, treated as OR
gbif_has_distribution("145953242",
                      country = c("NL","BE"),
                      status = c("PRESENT", "DOUBTFUL"))

#> [1] TRUE

# use alternative names: countryCode, occurrenceStatus.
# Function works. Warning is given.
gbif_has_distribution("145953242", countryCode = c("NL","BE"),
                 occurrenceStatus = c("PRESENT", "DOUBTFUL"))

#> Warning: countryCode renamed to country

#> Warning: occurrenceStatus renamed to status

#> [1] TRUE

# Case insensitive
gbif_has_distribution("145953242", countryCode = "be",
                 status = "PRESENT",
                 establishmentMeans = "InTrOdUcEd")

#> Warning: countryCode renamed to country

#> [1] TRUE