Pathway count indicator figure

Function to get number of taxa introduced by different pathways. Possible breakpoint: kingdom.

get_table_pathways(df, category = NULL, n_species = 5,
  kingdom = "kingdom")

Arguments

df
df.

category
NULL or character. One of the kingdoms as given in GBIF:

"Plantae"

"Animalia"

"Fungi"

"Chromista"

"Archaea"

"Bacteria"

"Protozoa"

"Viruses"

"incertae sedis"

It can also be one of the following not kingdoms: #'

Not Chordata. Default: NULL.

n_species
integer The maximum number of species to return as examples per pathway. For groups with less species than n_species, all species are given. Default: 5.

kingdom
character. Name of the column of df containing information about kingdom. Default: "kingdom".

Value

a data.frame

Examples

library(readr)
data <- read_tsv("https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv")

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   key = col_double(),
#>   nubKey = col_double(),
#>   speciesKey = col_double(),
#>   acceptedKey = col_double(),
#>   first_observed = col_double(),
#>   last_observed = col_double()
#> )

#> See spec(...) for full column specifications.

#> Warning: 22 parsing failures.
#>   row            col               expected actual                                                                                                                     file
#> 12634 first_observed no trailing characters -06-08 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12634 last_observed  no trailing characters -06-12 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12635 first_observed no trailing characters -09-18 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12635 last_observed  no trailing characters -09-18 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12636 first_observed no trailing characters -10-13 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> ..... .............. ...................... ...... ........................................................................................................................
#> See problems(...) for more details.

get_table_pathways(data)

#> [1] "All column names present"

#> # A tibble: 20 x 4
#>    pathway_level1 pathway_level2       n examples                               
#>    <chr>          <chr>            <int> <chr>                                  
#>  1 unknown        ""                3579 Ensis directus, Erechtites glossantha,…
#>  2 contaminant    habitat_material    91 Oenothera subterminalis, Verbena bract…
#>  3 ""             nursery             21 Galinsoga quadriradiata, Galinsoga par…
#>  4 ""             on_animals         570 Lamarckia aurea, Alternanthera sessili…
#>  5 ""             on_plants            4 Goodyera repens, Moneses uniflora, Bot…
#>  6 ""             seed               618 Cyperus eragrostis, Eleusine africana,…
#>  7 ""             timber              10 Ononis natrix, Melampyrum nemorosum, E…
#>  8 corridor       water               22 Lithoglyphus naticoides, Limnomysis be…
#>  9 escape         agriculture         85 Sinapis alba, Hyssopus officinalis, Av…
#> 10 ""             aquaculture         19 Cyprinus carpio, Orconectes limosus, P…
#> 11 ""             food_bait           29 Melilotus sulcatus, Ploceus melanoceph…
#> 12 ""             horticulture      1066 Solanum pseudocapsicum, Geum macrophyl…
#> 13 ""             ornamental           7 Lepomis gibbosus, Acipenser ruthenus, …
#> 14 ""             pet                  5 Orconectes limosus, Ferrissia californ…
#> 15 natural        dispersal           65 Mya arenaria, Sander lucioperca, Helia…
#> 16 release        biological_cont…     3 Hypophthalmichthys molitrix, Ctenophar…
#> 17 stowaway       ballast_water       34 Hemigrapsus takanoi, Ensis directus, P…
#> 18 ""             hull_fouling        20 Austrominius modestus, Amphibalanus im…
#> 19 ""             other               11 Proasellus coxalis, Girardia tigrina, …
#> 20 ""             people_luggage      10 Erodium moschatum, Trifolium nigrescen…

# Specify kingdom
get_table_pathways(data, "Plantae")

#> [1] "All column names present"

#> # A tibble: 12 x 4
#>    pathway_level1 pathway_level2      n examples                                
#>    <chr>          <chr>           <int> <chr>                                   
#>  1 unknown        ""               2407 Symphyotrichum lanceolatum, Euclidium s…
#>  2 contaminant    habitat_materi…    91 Bothriochloa ischaemum, Geranium purpur…
#>  3 ""             nursery            21 Euphorbia hypericifolia, Cardamine cory…
#>  4 ""             on_animals        565 Silene conoidea, Mesembryanthemum nodif…
#>  5 ""             on_plants           4 Orthilia secunda, Goodyera repens, Mone…
#>  6 ""             seed              618 Myagrum perfoliatum, Lathyrus ochrus, R…
#>  7 ""             timber             10 Carex muricata, Ononis natrix, Impatien…
#>  8 escape         agriculture        85 Phaseolus coccineus, Lactuca sativa, An…
#>  9 ""             food_bait          22 Persea americana, Cyperus rotundus, Pas…
#> 10 ""             horticulture     1066 Ranunculus lanuginosus, Gilia tricolor,…
#> 11 natural        dispersal          36 Reynoutria sachalinensis, Helianthus tu…
#> 12 stowaway       people_luggage     10 Soliva sessilis, Medicago littoralis, P…

# Sepcify number of species to include in examples
get_table_pathways(data, "Plantae", n_species = 8)

#> [1] "All column names present"

#> # A tibble: 12 x 4
#>    pathway_level1 pathway_level2      n examples                                
#>    <chr>          <chr>           <int> <chr>                                   
#>  1 unknown        ""               2407 Arabis auriculata, Onobrychis viciifoli…
#>  2 contaminant    habitat_materi…    91 Oenothera moravica, Odontites luteus, E…
#>  3 ""             nursery            21 Axonopus fissifolius, Wolffia columbian…
#>  4 ""             on_animals        565 Triraphis mollis, Senecio pterophorus, …
#>  5 ""             on_plants           4 Goodyera repens, Moneses uniflora, Orth…
#>  6 ""             seed              618 Erodium malacoides, Polygonum achoreum,…
#>  7 ""             timber             10 Euphorbia vermiculata, Epilobium ciliat…
#>  8 escape         agriculture        85 Nicotiana rustica, Cicer arietinum, Len…
#>  9 ""             food_bait          22 Solanum pimpinellifolium, Capsicum annu…
#> 10 ""             horticulture     1066 Phytolacca americana, Melica altissima,…
#> 11 natural        dispersal          36 Elodea canadensis, Prunus serotina, Elo…
#> 12 stowaway       people_luggage     10 Cotula australis, Trifolium nigrescens,…