Update the list with all downloads

This function opens a (tab-separated) text file containing all occurrence downloads from GBIF and updates the status of all downloads with status RUNNING or PREPARING. If the specifid download is not present it will be add.

update_download_list(file, download_to_add, input_checklist,
  url_doi_base = "https://doi.org/")

Arguments

file
text file containing all occurrence downloads from GBIF.

download_to_add
character. A GBIF download key to be added to file.

input_checklist
text file with taxon keys whose occurrences you want to download.

url_doi_base
character. doi base URL; url_doi_base + doi form a link to a page with download information. Default: "https://doi.org/".

Value

message with the performed updates

Details

If a download key is passed which is not present in the file it will be added as a new line.