Apply GAM to time series and assess emerging status

This function applies generalized additive models (GAM) to assess emerging status for a certain time window.

apply_gam(
  df,
  y_var,
  eval_years,
  year = "year",
  taxonKey = "taxonKey",
  type_indicator = "observations",
  baseline_var = NULL,
  p_max = 0.1,
  taxon_key = NULL,
  name = NULL,
  df_title = NULL,
  x_label = "year",
  y_label = "Observations",
  saveplot = FALSE,
  dir_name = NULL,
  verbose = FALSE
)

Arguments

df
df. A dataframe containing temporal data.

y_var
character. Name of column containing variable to model. It has to be passed as string, e.g. "occurrences".

eval_years
numeric. Temporal value(s) when emerging status has to be evaluated.

year
character. Name of column containing temporal values. It has to be passed as string, e.g. "time". Default: "year".

taxonKey
character. Name of column containing taxon IDs. It has to be passed as string, e.g. "taxon". Default: "taxonKey".

type_indicator
character. One of "observations", "occupancy". Used in title of the output plot. Default: "observations".

baseline_var
character. Name of the column containing values to use as additional covariate. Such covariate is introduced in the model to correct research effort bias. Default: NULL. If NULL internal variable method_em = "basic", otherwise method_em = "correct_baseline". Value of method_em will be part of title of output plot.

p_max
numeric. A value between 0 and 1. Default: 0.1.

taxon_key
numeric, character. Taxon key the timeseries belongs to. Used exclusively in graph title and filename (if saveplot = TRUE). Default: NULL.

name
character. Species name the timeseries belongs to. Used exclusively in graph title and filename (if saveplot = TRUE). Default: NULL.

df_title
character. Any string you would like to add to graph titles and filenames (if saveplot = TRUE). The title is always composed of: "GAM" + type_indicator + method_em + taxon_key

name + df_title separated by underscore ("_"). Default: NULL.

x_label
character. x-axis label of output plot. Default: "year".

y_label
character. y-axis label of output plot. Default: "number of observations".

saveplot
logical. If TRUE the plots are authomatically saved. Default: FALSE.

dir_name
character. Path of directory where saving plots. If path doesn't exists, directory will be created. Example: "./output/graphs/". If NULL, plots are saved in current directory. Default: NULL.

verbose
logical. If TRUE status of processing and possible issues are returned. Default: FALSE.

Value

list. List with six objects:

model: gam object. The model as returned by GAM.

first_derivative: df. Data.frame with details of first derivatives. It contains following columns:

var: character. Column name the smoother is applied to.

data: numeric. Data in columns defined by var.

derivative: numeric. Value of first derivative.

se: numeric. Standard error of derivative.

crit: numeric. Critical value required such that derivative + (se * crit) and derivative - (se * crit) form the upper and lower bounds of the confidence interval on the first derivative of the estimated smooth at the specific confidence level. In our case the confidence level is hard-coded: 0.8. Then crit <- qnorm(p = (1-0.8)/2, mean = 0, sd = 1, lower.tail = FALSE).

lower: numeric. Lower bound of the confidence interval of the estimated smooth.

upper: numeric. Upper bound of the confidence interval of the estimated smooth.

second_derivative: df. Data.frame with details of second derivatives. Same columns as first_derivatives.

plot: a ggplot2 object. Plot of GAM output and emerging. NULL if GAM cannot be applied.

Examples

if (FALSE) {
df_gam <- tibble(
  taxonKey = rep(3003709, 24),
  canonicalName = rep("Rosa glauca", 24),
  year = seq(1995, 2018),
  n = c(
    1, 1, 0, 0, 0, 2, 0, 0, 1, 3, 1, 2, 0, 5, 0, 5, 4, 2, 1,
    1, 3, 3, 8, 10
  ),
  n_class = c(
    229, 555, 1116, 939, 919, 853, 442, 532, 623, 1178, 732, 371, 1053,
    1001, 1550, 1142, 1076, 1310, 922, 1773, 1637,
    1866, 2234, 2013
  )
)
# apply GAM to n without baseline as covariate
apply_gam(df_gam,
  y_var = "n",
  eval_years = 2018,
  taxon_key = 3003709,
  name = "Rosa glauca",
  verbose = TRUE
)
# apply GAM using baseline data in column n_class as covariate
apply_gam(df_gam,
  y_var = "n",
  eval_years = 2018,
  baseline_var = "n_class",
  taxon_key = 3003709,
  name = "Rosa glauca",
  verbose = TRUE
)
# apply GAM using n as occupancy values, evaluate on two years. No baseline
apply_gam(df_gam,
  y_var = "n",
  eval_years = c(2017, 2018),
  taxon_key = 3003709,
  type_indicator = "occupancy",
  name = "Rosa glauca",
  y_label = "occupancy",
  verbose = TRUE
)
# apply GAM using n as occupancy values and n_class as covariate (baseline)
apply_gam(df_gam,
  y_var = "n",
  eval_years = c(2017, 2018),
  baseline_var = "n_class",
  taxon_key = 3003709,
  type_indicator = "occupancy",
  name = "Rosa glauca",
  y_label = "occupancy",
  verbose = TRUE
)
# How to use other arguments
apply_gam(df_gam,
  y_var = "n",
  eval_years = c(2017, 2018),
  baseline_var = "n_class",
  p_max = 0.3,
  taxon_key = "3003709",
  type_indicator = "occupancy",
  name = "Rosa glauca",
  df_title = "Belgium",
  x_label = "time (years)",
  y_label = "area of occupancy (km2)",
  saveplot = TRUE,
  dir_name = "./data/",
  verbose = TRUE
)
}