Pathway count indicator figure

Function to get number of taxa introduced by different pathways. Possible breakpoint: kingdom.

get_table_pathways(df, category = NULL, n_species = 5,
  kingdom = "kingdom")

Arguments

df
df.

category
NULL or character. One of the kingdoms as given in GBIF:

"Plantae"

"Animalia"

"Fungi"

"Chromista"

"Archaea"

"Bacteria"

"Protozoa"

"Viruses"

"incertae sedis"

It can also be one of the following not kingdoms: #'

Not Chordata. Default: NULL.

n_species
integer The maximum number of species to return as examples per pathway. For groups with less species than n_species, all species are given. Default: 5.

kingdom
character. Name of the column of df containing information about kingdom. Default: "kingdom".

Value

a data.frame

Examples

library(readr)

#> Warning: package 'readr' was built under R version 3.5.2

data <- read_tsv("https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv")

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   key = col_double(),
#>   nubKey = col_double(),
#>   speciesKey = col_double(),
#>   acceptedKey = col_double(),
#>   first_observed = col_double(),
#>   last_observed = col_double(),
#>   `invasion stage` = col_logical(),
#>   habitat = col_logical()
#> )

#> See spec(...) for full column specifications.

#> Warning: 307 parsing failures.
#>   row            col           expected      actual                                                                                                                     file
#> 12205 invasion stage 1/0/T/F/TRUE/FALSE established 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12205 habitat        1/0/T/F/TRUE/FALSE freshwater  'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12206 invasion stage 1/0/T/F/TRUE/FALSE established 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12206 habitat        1/0/T/F/TRUE/FALSE freshwater  'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12207 invasion stage 1/0/T/F/TRUE/FALSE established 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> ..... .............. .................. ........... ........................................................................................................................
#> See problems(...) for more details.

get_table_pathways(data)

#> [1] "All column names present"

#> # A tibble: 19 x 4
#>    pathway_level1 pathway_level2       n examples                               
#>    <chr>          <chr>            <int> <chr>                                  
#>  1 contaminant    habitat_material    91 Silene italica, Oenothera angustissima~
#>  2 ""             nursery             21 Schismus barbatus, Phyllanthus tenellu~
#>  3 ""             on_animals         569 Panicum bulbosum, Aegilops cylindrica,~
#>  4 ""             on_plants            4 Goodyera repens, Moneses uniflora, Ort~
#>  5 ""             seed               616 Rumex pseudonatronatus, Scolymus macul~
#>  6 ""             timber              10 Silene baccifera, Epilobium ciliatum, ~
#>  7 corridor       water               21 Orchestia cavimana, Proasellus meridia~
#>  8 escape         agriculture         84 Helianthus tuberosus, Brassica napus, ~
#>  9 ""             aquaculture         19 Pontastacus leptodactylus, Cyprinus ca~
#> 10 ""             food_bait           28 Solanum pimpinellifolium, Cucurbita pe~
#> 11 ""             horticulture      1072 Cotoneaster sternianus, Rosa virginian~
#> 12 ""             ornamental           7 Umbra pygmaea, Lepomis gibbosus, Acipe~
#> 13 ""             pet                  5 Hippolyte desmarestii, Pontastacus lep~
#> 14 release        biological_cont~     3 Ctenopharyngodon idella, Hypophthalmic~
#> 15 stowaway       ballast_water       34 Tasserkidrilus americanus, Mercenaria ~
#> 16 ""             hull_fouling        20 Austrominius modestus, Balanus tintinn~
#> 17 ""             other               11 Proasellus meridianus, Teredo navalis,~
#> 18 ""             people_luggage      10 Trifolium suffocatum, Paronychia argen~
#> 19 unknown        ""                 212 Sisyrinchium montanum, Lindernia dubia~

# Specify kingdom
get_table_pathways(data, "Plantae")

#> [1] "All column names present"

#> # A tibble: 11 x 4
#>    pathway_level1 pathway_level2      n examples                                
#>    <chr>          <chr>           <int> <chr>                                   
#>  1 contaminant    habitat_materi~    91 Parapholis incurva, Oenothera moravica,~
#>  2 ""             nursery            21 Linaria triornithophora, Cuscuta suaveo~
#>  3 ""             on_animals        564 Sporobolus creber, Malva nicaeensis, Xa~
#>  4 ""             on_plants           4 Moneses uniflora, Orthilia secunda, Bot~
#>  5 ""             seed              616 Chenopodiastrum simplex, Malva maroccan~
#>  6 ""             timber             10 Impatiens parviflora, Carex loliacea, M~
#>  7 escape         agriculture        84 Scorzonera hispanica, Hyssopus officina~
#>  8 ""             food_bait          22 Solanum pimpinellifolium, Passiflora ed~
#>  9 ""             horticulture     1072 Onopordum beckianum, Amaranthus cruentu~
#> 10 stowaway       people_luggage     10 Paronychia argentea, Medicago littorali~
#> 11 unknown        ""                210 Sabulina mediterranea, Malva punctata, ~

# Sepcify number of species to include in examples
get_table_pathways(data, "Plantae", n_species = 8)

#> [1] "All column names present"

#> # A tibble: 11 x 4
#>    pathway_level1 pathway_level2      n examples                                
#>    <chr>          <chr>           <int> <chr>                                   
#>  1 contaminant    habitat_materi~    91 Ornithopus compressus, Centaurea aspera~
#>  2 ""             nursery            21 Symphyotrichum subulatum, Schismus barb~
#>  3 ""             on_animals        564 Panicum schinzii, Sporobolus elongatus,~
#>  4 ""             on_plants           4 Orthilia secunda, Goodyera repens, Botr~
#>  5 ""             seed              616 Bromus squarrosus, Silene nocturna, Set~
#>  6 ""             timber             10 Symphyotrichum pilosum, Symphyotrichum ~
#>  7 escape         agriculture        84 Lepidium sativum, Allium schoenoprasum,~
#>  8 ""             food_bait          22 Ficus carica, Cucurbita maxima, Thladia~
#>  9 ""             horticulture     1072 Allium moly, Bergenia crassifolia, Pulm~
#> 10 stowaway       people_luggage     10 Ranunculus muricatus, Parentucellia lat~
#> 11 unknown        ""                210 Rumex thyrsiflorus, Asperula laevigata,~