Pathway count indicator figure

Function to get number of taxa introduced by different pathways. Possible breakpoint: kingdom.

get_table_pathways(df, category = NULL, n_species = 5,
  kingdom = "kingdom")

Arguments

df
df.

category
NULL or character. One of the kingdoms as given in GBIF:

"Plantae"

"Animalia"

"Fungi"

"Chromista"

"Archaea"

"Bacteria"

"Protozoa"

"Viruses"

"incertae sedis"

It can also be one of the following not kingdoms: #'

Not Chordata. Default: NULL.

n_species
integer The maximum number of species to return as examples per pathway. For groups with less species than n_species, all species are given. Default: 5.

kingdom
character. Name of the column of df containing information about kingdom. Default: "kingdom".

Value

a data.frame

Examples

library(readr)
datafile <- paste0(
  "https://raw.githubusercontent.com/trias-project/pipeline/master/data/",
  "interim/test_data_output_checklist_indicators.tsv"
)
data <- read_tsv(datafile)

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   key = col_double(),
#>   nubKey = col_double(),
#>   speciesKey = col_double(),
#>   acceptedKey = col_double(),
#>   first_observed = col_double(),
#>   last_observed = col_double()
#> )

#> See spec(...) for full column specifications.

#> Warning: 22 parsing failures.
#>   row            col               expected actual                                                                                                                     file
#> 12634 first_observed no trailing characters -06-08 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12634 last_observed  no trailing characters -06-12 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12635 first_observed no trailing characters -09-18 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12635 last_observed  no trailing characters -09-18 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> 12636 first_observed no trailing characters -10-13 'https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv'
#> ..... .............. ...................... ...... ........................................................................................................................
#> See problems(...) for more details.

get_table_pathways(data)

#> [1] "All column names present"

#> # A tibble: 20 x 4
#>    pathway_level1 pathway_level2       n examples                               
#>    <chr>          <chr>            <int> <chr>                                  
#>  1 unknown        ""                3579 Contarinia pisi, Hedera colchica, Bora…
#>  2 contaminant    habitat_material    91 Ambrosia psilostachya, Centaurea asper…
#>  3 ""             nursery             21 Cardamine corymbosa, Acalypha indica, …
#>  4 ""             on_animals         570 Eragrostis curvula, Tribulus terrestri…
#>  5 ""             on_plants            4 Orthilia secunda, Goodyera repens, Bot…
#>  6 ""             seed               618 Silene conoidea, Isatis tinctoria, Lag…
#>  7 ""             timber              10 Euphorbia vermiculata, Ononis natrix, …
#>  8 corridor       water               22 Lithoglyphus naticoides, Limnomysis be…
#>  9 escape         agriculture         85 Triticum monococcum, Phacelia tanaceti…
#> 10 ""             aquaculture         19 Orconectes limosus, Pontastacus leptod…
#> 11 ""             food_bait           29 Allium ramosum, Cucurbita maxima, Pass…
#> 12 ""             horticulture      1066 Corynabutilon suntense, Thalictrum spe…
#> 13 ""             ornamental           7 Carassius auratus, Acipenser gueldenst…
#> 14 ""             pet                  5 Hippolyte desmarestii, Orconectes limo…
#> 15 natural        dispersal           65 Mercenaria mercenaria, Dreissena polym…
#> 16 release        biological_cont…     3 Hypophthalmichthys molitrix, Ctenophar…
#> 17 stowaway       ballast_water       34 Mercenaria mercenaria, Melita nitida, …
#> 18 ""             hull_fouling        20 Ficopomatus enigmaticus, Proasellus me…
#> 19 ""             other               11 Girardia tigrina, Proasellus coxalis, …
#> 20 ""             people_luggage      10 Cotula australis, Poa infirma, Trifoli…

# Specify kingdom
get_table_pathways(data, "Plantae")

#> [1] "All column names present"

#> # A tibble: 12 x 4
#>    pathway_level1 pathway_level2      n examples                                
#>    <chr>          <chr>           <int> <chr>                                   
#>  1 unknown        ""               2407 Melampodium montanum, Chorispora tenell…
#>  2 contaminant    habitat_materi…    91 Erigeron canadensis, Verbascum virgatum…
#>  3 ""             nursery            21 Gamochaeta pensylvanica, Pilea microphy…
#>  4 ""             on_animals        565 Eragrostis alveiformis, Medicago lacini…
#>  5 ""             on_plants           4 Orthilia secunda, Moneses uniflora, Bot…
#>  6 ""             seed              618 Lotus maritimus, Ranunculus marginatus,…
#>  7 ""             timber             10 Euphorbia vermiculata, Carex loliacea, …
#>  8 escape         agriculture        85 Phaseolus coccineus, Allium schoenopras…
#>  9 ""             food_bait          22 Capsicum annuum, Passiflora edulis, Phy…
#> 10 ""             horticulture     1066 Santolina chamaecyparissus, Catalpa eru…
#> 11 natural        dispersal          36 Echinocystis lobata, Bunias orientalis,…
#> 12 stowaway       people_luggage     10 Soliva sessilis, Cotula australis, Pare…

# Sepcify number of species to include in examples
get_table_pathways(data, "Plantae", n_species = 8)

#> [1] "All column names present"

#> # A tibble: 12 x 4
#>    pathway_level1 pathway_level2      n examples                                
#>    <chr>          <chr>           <int> <chr>                                   
#>  1 unknown        ""               2407 Cyperus longus, Veronica cymbalaria, Pi…
#>  2 contaminant    habitat_materi…    91 Sagina alexandrae, Oenothera cambrica, …
#>  3 ""             nursery            21 Portulaca nitida, Gamochaeta pensylvani…
#>  4 ""             on_animals        565 Ambrosia trifida, Cenchrus longispinus,…
#>  5 ""             on_plants           4 Orthilia secunda, Moneses uniflora, Bot…
#>  6 ""             seed              618 Neslia paniculata, Festuca valesiaca, L…
#>  7 ""             timber             10 Symphyotrichum pilosum, Carex muricata,…
#>  8 escape         agriculture        85 Allium cepa, Triticum turgidum, Avena s…
#>  9 ""             food_bait          22 Physalis peruviana, Persea americana, P…
#> 10 ""             horticulture     1066 Gypsophila scorzonerifolia, Lathyrus ni…
#> 11 natural        dispersal          36 Berberis aquifolium, Sporobolus anglicu…
#> 12 stowaway       people_luggage     10 Soliva sessilis, Trifolium suffocatum, …